Updated a namspace and moved some interface definitions into their own files
