+++ Assets/MixedRealityToolkit/InputSystem/Editor/SpeechInputHandlerEditor.cs
@@ -3,7 +3,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.MixedReality.Toolkit.InputSystem.InputSources;
+using Microsoft.MixedReality.Toolkit.InputSystem.Sources;
 using Microsoft.MixedReality.Toolkit.InputSystem.Utilities.Interactions;
 using UnityEditor;
 using UnityEngine;
