+++ Assets/MixedRealityToolkit/InputSystem/InputEventData/BaseInputEventData.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.InputSystem.InputSources;
+using Microsoft.MixedReality.Toolkit.InputSystem.Sources;
 using UnityEngine.EventSystems;
 
 namespace Microsoft.MixedReality.Toolkit.InputSystem.EventData
