+++ Assets/MixedRealityToolkit/InputSystem/InputEventData/SpeechEventData.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
-using Microsoft.MixedReality.Toolkit.InputSystem.InputSources;
+using Microsoft.MixedReality.Toolkit.InputSystem.Sources;
 using UnityEngine.EventSystems;
 
 #if UNITY_WSA || UNITY_STANDALONE_WIN
