+++ Assets/MixedRealityToolkit/InputSystem/Sources/GenericInputSource.cs
@@ -10,7 +10,7 @@ using Microsoft.MixedReality.Toolkit.Internal.Interfaces;
 using Microsoft.MixedReality.Toolkit.Internal.Managers;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.InputSystem.InputSources
+namespace Microsoft.MixedReality.Toolkit.InputSystem.Sources
 {
     /// <summary>
     /// Base class for input sources that don't inherit from MonoBehaviour.
