+++ Assets/MixedRealityToolkit/InputSystem/Sources/GenericInputSources.cs
@@ -8,7 +8,7 @@ using Microsoft.MixedReality.Toolkit.Internal.Interfaces;
 using Microsoft.MixedReality.Toolkit.Internal.Managers;
 using UnityEngine;
 
-namespace Microsoft.MixedReality.Toolkit.InputSystem.InputSources
+namespace Microsoft.MixedReality.Toolkit.InputSystem.Sources
 {
     /// <summary>
     /// This class manages the Input Sources recognized by <see cref="Input.GetJoystickNames"/>.
