+++ Assets/MixedRealityToolkit/_Core/Interfaces/IFocusProvider.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b439356d3f6f439abeb9e06c8cfbe39c
+timeCreated: 1523919204
