+++ Assets/MixedRealityToolkit/_Core/Interfaces/IGazeProvider.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using Microsoft.MixedReality.Toolkit.InputSystem.Sources;
+
+namespace Microsoft.MixedReality.Toolkit.Internal.Interfaces
+{
+    public interface IGazeProvider : IInputSource { }
+}
