+++ Assets/MixedRealityToolkit/_Core/Interfaces/IGazeProvider.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f1eb327b35384be890e24f8998a1cc1e
+timeCreated: 1523919204
