--- Assets/MixedRealityToolkit/_Core/Interfaces/InputSystem.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ebf5d53ad0c93444bb66cab825d29f45
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
